rootcause::hooks::formatting_overrides::attachment

Trait AttachmentFormattingOverride

pub trait AttachmentFormattingOverride<A>:
    'static
    + Send
    + Sync {
    // Provided methods
    fn display(
        &self,
        attachment: ReportAttachmentRef<'_, A>,
        attachment_parent: Option<AttachmentParent<'_>>,
        formatter: &mut Formatter<'_>,
    ) -> Result { ... }
    fn display_preformatted(
        &self,
        attachment: ReportAttachmentRef<'_, PreformattedAttachment>,
        attachment_parent: Option<AttachmentParent<'_>>,
        formatter: &mut Formatter<'_>,
    ) -> Result { ... }
    fn debug(
        &self,
        attachment: ReportAttachmentRef<'_, A>,
        attachment_parent: Option<AttachmentParent<'_>>,
        formatter: &mut Formatter<'_>,
    ) -> Result { ... }
    fn debug_preformatted(
        &self,
        attachment: ReportAttachmentRef<'_, PreformattedAttachment>,
        attachment_parent: Option<AttachmentParent<'_>>,
        formatter: &mut Formatter<'_>,
    ) -> Result { ... }
    fn preferred_formatting_style(
        &self,
        attachment: ReportAttachmentRef<'_, dyn Any>,
        report_formatting_function: FormattingFunction,
    ) -> AttachmentFormattingStyle { ... }
}

Trait for customizing how attachments of a specific type are formatted in error reports.

This trait allows you to override the default formatting behavior for attachments of type A. You can customize both Display and Debug formatting, as well as handle preformatted attachments and specify preferred formatting styles.

Type Parameters

A - The attachment type that this formatter handles

Default Implementations

All methods have default implementations that delegate to the unhooked formatting, so you only need to override the methods for the formatting you want to customize.

Examples

Basic custom Display formatting:

use core::fmt;

use rootcause::{
    hooks::formatting_overrides::attachment::{AttachmentFormattingOverride, AttachmentParent},
    report_attachment::ReportAttachmentRef,
};

struct MyData {
    value: i32,
    name: String,
}

struct MyDataFormatter;

impl AttachmentFormattingOverride<MyData> for MyDataFormatter {
    fn display(
        &self,
        attachment: ReportAttachmentRef<'_, MyData>,
        _parent: Option<AttachmentParent<'_>>,
        f: &mut fmt::Formatter<'_>,
    ) -> fmt::Result {
        let data = attachment.inner();
        write!(f, "{}: {}", data.name, data.value)
    }
}

Provided Methods

fn display( &self, attachment: ReportAttachmentRef<'_, A>, attachment_parent: Option<AttachmentParent<'_>>, formatter: &mut Formatter<'_>, ) -> Result

Formats the attachment using Display formatting.

This method is called when the attachment needs to be displayed in a user-friendly format. The default implementation delegates to the attachment’s unhooked Display formatting.

Arguments

attachment - Reference to the attachment being formatted

attachment_parent - Optional context about the parent report

formatter - The formatter to write output to

Examples

use core::fmt;

use rootcause::{
    hooks::formatting_overrides::attachment::{AttachmentFormattingOverride, AttachmentParent},
    report_attachment::ReportAttachmentRef,
};

struct ErrorCode(u32);
struct ErrorCodeFormatter;

impl AttachmentFormattingOverride<ErrorCode> for ErrorCodeFormatter {
    fn display(
        &self,
        attachment: ReportAttachmentRef<'_, ErrorCode>,
        _parent: Option<AttachmentParent<'_>>,
        f: &mut fmt::Formatter<'_>,
    ) -> fmt::Result {
        write!(f, "Error Code: 0x{:04X}", attachment.inner().0)
    }
}

fn display_preformatted( &self, attachment: ReportAttachmentRef<'_, PreformattedAttachment>, attachment_parent: Option<AttachmentParent<'_>>, formatter: &mut Formatter<'_>, ) -> Result

Formats a preformatted attachment using Display formatting.

This method handles attachments that have been preformatted (typically done using ReportRef::preformat). The default implementation delegates to the attachment’s unhooked Display formatting.

Arguments

attachment - Reference to the preformatted attachment

attachment_parent - Optional context about the parent report

formatter - The formatter to write output to

fn debug( &self, attachment: ReportAttachmentRef<'_, A>, attachment_parent: Option<AttachmentParent<'_>>, formatter: &mut Formatter<'_>, ) -> Result

Formats the attachment using Debug formatting.

This method is called when the attachment needs to be displayed in a debug-friendly format. The default implementation delegates to the attachment’s unhooked Debug formatting.

Arguments

attachment - Reference to the attachment being formatted

attachment_parent - Optional context about the parent report

formatter - The formatter to write output to

fn debug_preformatted( &self, attachment: ReportAttachmentRef<'_, PreformattedAttachment>, attachment_parent: Option<AttachmentParent<'_>>, formatter: &mut Formatter<'_>, ) -> Result

Formats a preformatted attachment using Debug formatting.

This method handles attachments that have been preformatted (typically done using ReportRef::preformat). The default implementation delegates to the attachment’s unhooked Debug formatting.

Arguments

attachment - Reference to the preformatted attachment

attachment_parent - Optional context about the parent report

formatter - The formatter to write output to

fn preferred_formatting_style( &self, attachment: ReportAttachmentRef<'_, dyn Any>, report_formatting_function: FormattingFunction, ) -> AttachmentFormattingStyle

Determines the preferred formatting style for this attachment.

This method allows the formatter to specify how the attachment should be presented in the overall report structure (inline, with header, in appendix, etc.). The default implementation delegates to the attachment’s unhooked preference.

Arguments

attachment - Reference to the attachment (as dyn Any as it can be either A or a PreformattedAttachment)

report_formatting_function - Whether the overall report uses Display or Debug formatting