rootcause::hooks::report_creation

Function register_attachment_collector_hook

pub fn register_attachment_collector_hook<A, C>(collector: C)
where
    A: 'static + Send + Sync,
    C: AttachmentCollectorHook<A> + Send + Sync + 'static,

Registers an attachment collector hook that will automatically collect and attach data to reports.

Once registered, the collector will be automatically invoked for every report that gets created. The collected data will be formatted using the collector’s associated handler and attached to the report.

Note that internally this function converts the attachment collector into a report creation hook and registers it using the same system as register_report_creation_hook.

Registration Order

Attachment collectors are called in the order they were registered, after any general report creation hooks have been executed.

Performance Considerations

Registered collectors will be called for every report creation, so they should be designed to be fast. Heavy operations in collectors can significantly impact the performance of error reporting throughout the application.

Examples

Registering a Custom Collector

use rootcause::{
    hooks::report_creation::{AttachmentCollectorHook, register_attachment_collector_hook},
    prelude::*,
};

struct MemoryUsageCollector;
#[derive(Debug)]
struct MemoryInfo {
    used_mb: u64,
}

impl AttachmentCollectorHook<MemoryInfo> for MemoryUsageCollector {
    type Handler = handlers::Debug;

    fn collect(&self) -> MemoryInfo {
        // This is a simplified example - in practice you'd use a proper method
        // to get memory usage information
        MemoryInfo { used_mb: 45 }
    }
}

// Register the collector - it will now collect memory info for all reports
register_attachment_collector_hook(MemoryUsageCollector);

Registering a Closure Collector

use rootcause::hooks::report_creation::register_attachment_collector_hook;

// Register a simple closure that collects the current thread ID as a string
register_attachment_collector_hook(|| {
    format!("Created on thread_id={:?}", std::thread::current().id())
});