rootcause::hooks::report_creation

Trait AttachmentCollectorHook

pub trait AttachmentCollectorHook<A>:
    'static
    + Send
    + Sync {
    type Handler: AttachmentHandler<A>;

    // Required method
    fn collect(&self) -> A;
}

A hook that collects data to be automatically attached to reports when they are created.

Attachment collector hooks provide a specialized way to automatically collect and attach specific types of data to all reports. Unlike ReportCreationHook, which provides full access to the report, attachment collectors are focused solely on gathering data to be attached.

Automatic Implementation

This trait is automatically implemented for any closure that returns a value implementing Display and Debug, using handlers::Display as the handler:

use rootcause::hooks::report_creation::register_attachment_collector_hook;

// This closure automatically implements AttachmentCollectorHook<String>
register_attachment_collector_hook(|| "timestamp".to_string());

Examples

Custom Collector Implementation

use rootcause::{
    hooks::report_creation::{AttachmentCollectorHook, register_attachment_collector_hook},
    prelude::*,
};

struct SystemInfoCollector;

impl AttachmentCollectorHook<String> for SystemInfoCollector {
    type Handler = handlers::Display;

    fn collect(&self) -> String {
        format!(
            "OS: {}, Arch: {}",
            std::env::consts::OS,
            std::env::consts::ARCH
        )
    }
}

// Register the collector
register_attachment_collector_hook(SystemInfoCollector);

Using a Closure

use rootcause::hooks::report_creation::register_attachment_collector_hook;

// Register a closure that collects the current working directory
register_attachment_collector_hook(|| {
    std::env::current_dir()
        .map(|p| p.display().to_string())
        .unwrap_or_else(|_| "unknown".to_string())
});

Required Associated Types

type Handler: AttachmentHandler<A>

The handler type used to format the collected data.

Required Methods

fn collect(&self) -> A

Collects the data to be attached to a report.

This method is called once for each report creation and should return the data that will be attached to the report. The data will be formatted using the associated Handler type.