rootcause::hooks::report_creation

Trait ReportCreationHook

pub trait ReportCreationHook:
    'static
    + Send
    + Sync {
    // Required methods
    fn on_local_creation(&self, report: ReportMut<'_, dyn Any, Local>);
    fn on_sendsync_creation(&self, report: ReportMut<'_, dyn Any, SendSync>);
}

A hook that is called whenever a report is created.

Report creation hooks provide a way to automatically modify or enhance reports as they are being created, without requiring changes to the code that creates the reports. This is useful for adding consistent metadata, logging, or performing other side effects.

If you only need to add attachments, then consider using an AttachmentCollectorHook instead, as it gives you an easier to use API for this use case.

Examples

use rootcause::{
    ReportMut,
    hooks::report_creation::{ReportCreationHook, register_report_creation_hook},
    markers::{Local, SendSync},
    prelude::*,
};

struct LoggingHook;

impl ReportCreationHook for LoggingHook {
    fn on_local_creation(&self, mut report: ReportMut<'_, dyn std::any::Any, Local>) {
        println!("Local report created: {}", report);
        let attachment = report_attachment!("Logged by LoggingHook");
        report.attachments_mut().push(attachment.into());
    }

    fn on_sendsync_creation(&self, mut report: ReportMut<'_, dyn std::any::Any, SendSync>) {
        println!("SendSync report created: {}", report);
        let attachment = report_attachment!("Logged by LoggingHook");
        report.attachments_mut().push(attachment.into());
    }
}

// Register the hook to activate it
register_report_creation_hook(LoggingHook);

Required Methods

fn on_local_creation(&self, report: ReportMut<'_, dyn Any, Local>)

Called when a Local report is created.

fn on_sendsync_creation(&self, report: ReportMut<'_, dyn Any, SendSync>)

Called when a SendSync report is created.